requirements.txt
uvicorn
fastapi
sqlalchemy
python-dotenv
pydantic
pydantic_settings
asyncpg
pydantic[email]
python-dateutil
python-jose
passlib
Werkzeug
python-multipart
python-keycloak==5.1.1